Fix/upgrade original/1 (#1)
* step 1

* Starting to improve the generated code (long way to go I think)

* Some clean up. Beginning to rework the code translation

* EOD check in

End of day check in

* .

* ..

* .3

* .5

* Parameter parsing is looking good (not done but good)

* .6

* .8 - breaking for now

* Neck deep. Recoding property generators

* Compiling after many changes including the beginnings of a more complete property generator (get / set)

* EOD - good progress on get/set generator and the different property types (IAmAProperty)

* SOD (Start of Day) handoff

* Properties starting to take shape (yay!)

* Looks pretty good for converting get/let/set properties. Almost there. Is working for a couple files. More testing is needed

* A little further. Starting to take shape. Long way to go but I'm surprised how far I've gotten

* Moving right along. Embedding different kind of blanket replacements, which is what the new Alice and Massage classes are. Onward and upward!

* eod

* SOD - Doing some clean up and renames

* More code cleanup. Moving toward a more onion like approach to the generated code instead of flat

* .

* End of session. get/set areas are looking a lot better. I think I know how to handle the  on error stuff but will tackle that later. It's amazing how just a few key functions can show such progress

* End of weekend. Created and started integrating the new CodeBlock class. It's not there, but it's definitely better

* Switching to CodeBlock class which is proving to be superior. I should now be able to accelerate upgrading the code as well as give myself a better look into the code when things go wrong

* EOD - Inching forward. Still integrating CodeBlock which is looking good

* SOD - just some coding before work. Added another (not yet working) unit test. I think I may handle corrections in translation as future unit tests

* SOD - Before hopefully slipping the new ICodeBlock/ICodeLine interfaces in place of AbstractCodeBlock

* EOD - Integrated ICodeLine, ICodeBlock, CodeLine, and CodeBlock throughout the incoming code.

* Inching

* Streamlining some names like CodeBlock -> Block

* Still integrating the CodeLine/CodeBlock stuff. Very close but I have to sleep

* 10 green tests. CodeBlock chain of objects appears to be running as intended. Now back to integrating/genereting all project. Hopefully the CodeBlock chain will simplify things a bit (especially indenation)

+++ Program.cs
@@ -1,32 +1,48 @@
-﻿using System.IO;
-using System.Linq;
+﻿using System.Collections.Generic;
+using System.IO;
 
-namespace VB2C
+namespace MetX.VB6ToCSharp
 {
     public static class Program
     {
-        public static void ConvertAll(string folderPath)
-        {
-            foreach (var fileToDelete in Directory.EnumerateFiles(folderPath))
-                File.Delete(fileToDelete);
-
-            var clsFiles = Directory.EnumerateFiles(@"I:\Sandy", "*.cls").ToList();
-            var frmFiles = Directory.EnumerateFiles(@"I:\Sandy", "*.frm").ToList();
-            var basFiles = Directory.EnumerateFiles(@"I:\Sandy", "*.bas").ToList();
-
-            foreach (var fileSet in new[] { clsFiles, basFiles, frmFiles })
-                foreach (var clsFile in fileSet)
-                {
-                    var ConvertObject = new ConvertCode();
-                    ConvertObject.ParseFile(clsFile, folderPath);
-                }
-        }
+        public const string SourceCodeFolderPath = @"I:\OneDrive\data\code\Slice and Dice\Sandy\";
+        public const string OutputFolderPath = @"I:\OneDrive\data\code\Slice and Dice\SandyC\";
+        // public const bool ClearOutputFolder = false;
 
         public static int Main(string[] args)
         {
-            ConvertAll(@"I:\SandyB\");
-
+            ConvertAllFiles();
             return 0;
+        }
+
+        public static void ConvertAllFiles()
+        {
+
+            /*
+            if(ClearOutputFolder)
+            {
+                // Delete previous run
+                foreach (var fileToDelete in Directory.EnumerateFiles(OutputFolderPath))
+                {
+                    File.SetAttributes(fileToDelete, FileAttributes.Normal);
+                    File.Delete(fileToDelete);
+                }
+            }
+
+            */
+            var fileSets = new List<IEnumerable<string>>
+            {
+                Directory.EnumerateFiles(SourceCodeFolderPath, "*.cls"),
+                Directory.EnumerateFiles(SourceCodeFolderPath, "*.frm"),
+                Directory.EnumerateFiles(SourceCodeFolderPath, "*.bas"),
+            };
+
+            foreach (var fileSet in fileSets)
+            foreach (var file in fileSet)
+            {
+                new ModuleConverter()
+                    .ConvertFile(file, OutputFolderPath);
+            }
         }
     }
 }
